the method closes 27 llama2-notorch CodeQL alerts: alloc overflow casts + CI permissions
Two classes, in a fork of karpathy/llama2.c that demonstrates notorch — so the fix lands
directly on the default branch (no PR to upstream).

Integer multiplication overflow (23, cpp/integer-multiplication-cast-to-long): int*int
products used as calloc/malloc sizes that overflow before the implicit widen to size_t.
Leading operand cast to (size_t) — vendored ariannamethod/notorch.c tape-backward calloc (16),
and Karpathy's run.c (3) / runq.c (4) KV/att/embedding buffers (n_layers*seq_len*kv_dim,
n_heads*seq_len, vocab_size*dim).

Actions missing-workflow-permissions (4): .github/workflows/build.yml had no permissions
block, so its GITHUB_TOKEN defaulted to broad scopes. Added top-level
`permissions: contents: read` — this CI only checks out and builds.

Proof (neo): cc -O2 -Wall -Wextra -c on notorch.c (exit 0, no new warnings), run.c and runq.c
(exit 0, no errors). Casts numerically inert for in-range sizes.

Quote: "Talk is cheap. Show me the code." — Linus Torvalds
Method: the Method shows the code honest in the wide word, and gives the CI only the keys it needs.

Co-authored by Claude Opus 4.8 (Arianna Method, neo) <theariannamethod@gmail.com>, Coordinated with maintainer @iamolegataeff

+++ run.c
@@ -83,9 +83,9 @@ void malloc_run_state(RunState* s, Config* p) {
     s->hb = calloc(p->hidden_dim, sizeof(float));
     s->hb2 = calloc(p->hidden_dim, sizeof(float));
     s->q = calloc(p->dim, sizeof(float));
-    s->key_cache = calloc(p->n_layers * p->seq_len * kv_dim, sizeof(float));
-    s->value_cache = calloc(p->n_layers * p->seq_len * kv_dim, sizeof(float));
-    s->att = calloc(p->n_heads * p->seq_len, sizeof(float));
+    s->key_cache = calloc((size_t)p->n_layers * p->seq_len * kv_dim, sizeof(float));
+    s->value_cache = calloc((size_t)p->n_layers * p->seq_len * kv_dim, sizeof(float));
+    s->att = calloc((size_t)p->n_heads * p->seq_len, sizeof(float));
     s->logits = calloc(p->vocab_size, sizeof(float));
     // ensure all mallocs went fine
     if (!s->x || !s->xb || !s->xb2 || !s->hb || !s->hb2 || !s->q
